# back/docs/swagger.yaml
basePath: /api/v1
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Game:
    properties:
      created_at:
        type: string
      id:
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Match:
    properties:
      created_at:
        type: string
      endTime:
        type: string
      id:
        type: integer
      matchPosition:
        type: integer
      playerOne:
        $ref: '#/definitions/models.User'
      playerOneID:
        type: integer
      playerTwo:
        $ref: '#/definitions/models.User'
      playerTwoID:
        type: integer
      scores:
        items:
          $ref: '#/definitions/models.Score'
        type: array
      startTime:
        type: string
      status:
        type: string
      tournament:
        $ref: '#/definitions/models.Tournament'
      tournamentID:
        type: integer
      tournamentStep:
        $ref: '#/definitions/models.TournamentStep'
      tournamentStepID:
        type: integer
      updated_at:
        type: string
      winner:
        $ref: '#/definitions/models.User'
      winnerID:
        type: integer
    required:
    - status
    type: object
  models.Media:
    properties:
      created_at:
        type: string
      file_extension:
        type: string
      file_name:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  models.NewUserPayload:
    properties:
      address:
        example: 1234 street
        type: string
      email:
        type: string
      password:
        example: password
        type: string
      phone:
        example: "1234567890"
        type: string
      username:
        example: username
        type: string
    type: object
  models.Score:
    properties:
      created_at:
        type: string
      id:
        type: integer
      match:
        $ref: '#/definitions/models.Match'
      matchID:
        type: integer
      player:
        $ref: '#/definitions/models.User'
      playerID:
        type: integer
      score:
        type: integer
      updated_at:
        type: string
    type: object
  models.Tag:
    properties:
      created_at:
        type: string
      games:
        items:
          $ref: '#/definitions/models.Game'
        type: array
      id:
        type: integer
      name:
        type: string
      tournaments:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
  models.Tournament:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      game:
        $ref: '#/definitions/models.Game'
      game_id:
        type: integer
      id:
        type: integer
      location:
        type: string
      maxPlayers:
        example: 32
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      name:
        type: string
      organizer_id:
        type: integer
      rounds:
        type: integer
      start_date:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tournament_steps:
        items:
          $ref: '#/definitions/models.TournamentStep'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - description
    - end_date
    - game_id
    - maxPlayers
    - name
    - rounds
    - start_date
    type: object
  models.TournamentStep:
    properties:
      created_at:
        type: string
      id:
        type: integer
      matches:
        items:
          $ref: '#/definitions/models.Match'
        type: array
      name:
        type: string
      sequence:
        type: integer
      tournamentID:
        type: integer
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      address:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      matches:
        items:
          $ref: '#/definitions/models.Match'
        type: array
      media:
        $ref: '#/definitions/models.Media'
      media_id:
        type: integer
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        type: string
      tournaments:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      updated_at:
        type: string
      username:
        type: string
    required:
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: Filter by isVerified
        in: query
        name: isVerified
        type: string
      - description: Sort by
        in: query
        name: sort
        type: string
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by country
        in: query
        name: country
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a user
      parameters:
      - description: User Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.User'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.NewUserPayload'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update a user
      parameters:
      - description: User Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.User'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - User
  /users/{id}/upload/picture:
    post:
      consumes:
      - multipart/form-data
      description: Upload a picture
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a picture
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"